- restructure
